+++ src/psykube/manifest/ingress/host/path.cr
@@ -1,8 +1,8 @@
 class Psykube::Manifest::Ingress::Host::Path
   Manifest.mapping({
-    port: String | UInt16,
+    port: String | Int32,
   })
 
-  def initialize(@port : String | UInt16)
+  def initialize(@port : String | Int32)
   end
 end
